+++ crates/utiles-core/src/lib.rs
@@ -24,9 +24,9 @@ pub mod errors;
 pub mod bbox;
 pub mod constants;
 pub mod fns;
-mod gdal;
+pub mod gdal;
 pub mod geojson;
-mod geostats;
+pub mod geostats;
 pub mod lint_error;
 pub mod lnglat;
 pub mod mbtiles;
@@ -44,7 +44,6 @@ pub mod tile_range;
 mod tile_tuple;
 pub mod tile_type;
 mod tilecrz;
-pub mod tilejson;
 pub mod traits;
 pub mod zoom;
