+++ crates/utiles/src/utilesqlite/mbtiles_async.rs
@@ -4,7 +4,8 @@ use deadpool_sqlite::{self, Config, Pool, PoolError, Runtime};
 use tilejson::TileJSON;
 use tracing::error;
 
-use utiles_core::mbtiles::metadata2tilejson;
+use crate::utilejson::metadata2tilejson;
+
 use utiles_core::mbtiles::metadata_row::MbtilesMetadataRow;
 use utiles_core::tile_data_row::TileData;
 use utiles_core::Tile;
