+++ utiles-pyo3/Cargo.toml
@@ -17,8 +17,6 @@ rusqlite.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 utiles = { path = "../crates/utiles" }
-utilesqlite = { path = "../crates/utilesqlite" }
-#utiles-cli = { path = "../crates/utiles" }
 
 [dev-dependencies]
 pyo3 = { workspace = true, features = ["auto-initialize"] }
